APERTA-8707 Search through Nested Questions instead of through tasks since Front Matter tasks may not exist though questions do

+++ lib/tasks/data-migrations/APERTA-8707-migrate-reviewer-report-task-questions-to-reviewer-report.rake
@@ -6,41 +6,31 @@ namespace :data do
         relevant_tasks = ['TahiStandardTasks::ReviewerReportTask', 'TahiStandardTasks::FrontMatterReviewerReportTask']
         task_count = Task.where(type: relevant_tasks).count
         STDOUT.puts("Task count: #{task_count}")
-        Task.where(type: relevant_tasks).find_each do |task|
-          STDOUT.puts("Processing #{task.title} id: #{task.id}")
-          reviewer_report_ids = []
-          task.nested_questions.each do |question|
-            STDOUT.puts("Processing Question #{question.id}")
-            if question.nested_question_answers.present?
-              question.nested_question_answers.find_each do |answer|
-                STDOUT.puts("Processing QuestionAnswer #{answer.id}")
-                reviewer_report = ReviewerReport.where(
-                  task: task,
-                  user: task.reviewer,
-                  decision: answer.decision
-                ).first_or_create!
-                unless reviewer_report_ids.include? reviewer_report.id
-                  reviewer_report_ids << reviewer_report.id
-                end
-                STDOUT.puts("Setting Answer Owner for answer: #{answer.id} for reviewer_report: #{reviewer_report.id}")
-                question.owner_type = ReviewerReport.name
-                answer.owner = reviewer_report
-                answer.save(validate: false)
-              end
-            else
-              if question.parent
-                question.owner_type = question.parent.owner_type
-              else
-                STDERR.puts("Found an orphan question without answers")
-                fail "Found an orphan question!"
-              end
+
+        reviewer_report_ids = []
+        NestedQuestion.where(owner_type: relevant_tasks).find_each do |question|
+          STDOUT.puts("Processing Question #{question.id}")
+          question.nested_question_answers.find_each do |answer|
+            STDOUT.puts("Processing QuestionAnswer #{answer.id}")
+            task = answer.owner
+            reviewer_report = ReviewerReport.where(
+              task: task,
+              user: task.reviewer,
+              decision: answer.decision
+            ).first_or_create!
+            answer.owner = reviewer_report
+            unless reviewer_report_ids.include? reviewer_report.id
+              reviewer_report_ids << reviewer_report.id
             end
-            question.save
+            STDOUT.puts("Setting Answer Owner for answer: #{answer.id} for reviewer_report: #{reviewer_report.id}")
+            answer.save(validate: false)
           end
-          STDOUT.puts("Task count: #{task_count}")
-          STDOUT.puts("Reviewer Report Count: #{reviewer_report_ids.length}")
+          question.update!(owner_type: ReviewerReport.name)
         end
-        if Task.where(type: relevant_tasks).map(&:nested_questions).flatten.empty?
+        STDOUT.puts("Task count: #{task_count}")
+        STDOUT.puts("Reviewer Report Count: #{reviewer_report_ids.length}")
+
+        if NestedQuestion.where(owner_type: relevant_tasks).empty?
           STDOUT.puts("Nested Question migration for Reviewer Report completed")
         else
           STDOUT.puts("Migration Failed, not all nested questions were moved")
@@ -48,17 +38,23 @@ namespace :data do
       end
 
       task reviewer_report_to_reviewer_report_task: :environment do
-        ReviewerReport.find_each do |report|
-          task = report.task
-
-          report.nested_questions.each do |question|
-            question.nested_question_answers.find_each do |answer|
-              answer.owner = task
-              answer.save(validate: false)
-            end
-            question.update!(owner_type: task.class.name)
+        # Not all Reviewer Report NestedQuestions are associated with a Reviewer Report
+        NestedQuestion.where(owner_type: ReviewerReport.name).find_each do |question|
+          task = question.owner.task
+          question.nested_question_answers.find_each do |answer|
+            answer.owner = task
+            answer.save(validate: false)
           end
-          report.destroy!
+          question.update!(owner_type: task.class.name)
+        end
+
+        ReviewerReport.each do |report|
+          if report.nested_questions.empty?
+            STDOUT.puts("Destroying ReviewerReport: #{report.id}")
+            report.destroy!
+          else
+            STDERR.puts("Error with a ReviewerReport (id: #{report.id}) that is still associated with nested questions")
+          end
         end
       end
     end
